+++ tcex/tcex_redis.py
@@ -16,12 +16,34 @@ class TcExRedis(object):
             port (string): The Redis port.
             rhash (string): The rhash value.
         """
-        self._hash = rhash
-        self._r = redis.StrictRedis(host=host, port=port)
+        self.hash = rhash
+        self.r = redis.StrictRedis(host=host, port=port)
+
+    def blpop(self, keys, timeout=30):
+        """RPOP a value off the first empty list in keys.
+
+        .. note:: If timeout is 0, the block indefinitely.
+
+        Args:
+            keys (string|list): The key(s) to pop the value.
+            timeout (int): The number of seconds to wait before blocking stops.
+
+        Returns:
+            (string): The response from Redis.
+        """
+        return self.r.blpop(keys, timeout)
 
     def create(self, key, value):
-        """Alias for hset method."""
-        self.hset(key, value)
+        """Create key/value pair in Redis.
+
+        Args:
+            key (string): The key to create in Redis.
+            value (any): The value to store in Redis.
+
+        Returns:
+            (string): The response from Redis.
+        """
+        return self.r.hset(self.hash, key, value)
 
     def delete(self, key):
         """Alias for hdel method."""
@@ -36,7 +58,7 @@ class TcExRedis(object):
         Returns:
             (string): The response from Redis.
         """
-        return self._r.hdel(self._hash, key)
+        return self.r.hdel(self.hash, key)
 
     def hget(self, key):
         """Read data from Redis for the provided key.
@@ -47,9 +69,9 @@ class TcExRedis(object):
         Returns:
             (any): The response data from Redis.
         """
-        data = self._r.hget(self._hash, key)
+        data = self.r.hget(self.hash, key)
         if data is not None and not isinstance(data, str):
-            data = str(self._r.hget(self._hash, key), 'utf-8')
+            data = str(self.r.hget(self.hash, key), 'utf-8')
         return data
 
     def hgetall(self):
@@ -61,7 +83,7 @@ class TcExRedis(object):
         Returns:
             (any): The response data from Redis.
         """
-        return self._r.hgetall(self._hash)
+        return self.r.hgetall(self.hash)
 
     def hset(self, key, value):
         """Create key/value pair in Redis.
@@ -73,8 +95,20 @@ class TcExRedis(object):
         Returns:
             (string): The response from Redis.
         """
-        return self._r.hset(self._hash, key, value)
+        return self.r.hset(self.hash, key, value)
 
     def read(self, key):
         """Alias for hget method."""
         return self.hget(key)
+
+    def rpush(self, name, values):
+        """Append/Push values to the end of list ``name``.
+
+        Args:
+            name (string): The channel/name of the list.
+            timeout (int): The number of seconds to wait before blocking stops.
+
+        Returns:
+            (string): The response from Redis.
+        """
+        return self.r.rpush(name, *values)